+++ test/sql/type.sql
@@ -44,5 +44,8 @@ select dlurlscheme(dlvalue('http://www.ljudmila.org/foo/bar/baz#123'));
 select dlurlpath(dlvalue('http://www.ljudmila.org/foo/bar/baz#123'));
 select dlurlpathonly(dlvalue('http://www.ljudmila.org/foo/bar/baz#123'));
 
+select dlurlpath(dlvalue('file:///foo/bar/baz#123'));
+select dlurlpathonly(dlvalue('file:///foo/bar/baz#123'));
+
 select dllinktype(dlvalue('http://www.ljudmila.org/foo/bar/baz#123'));
 select dllinktype(dlvalue('/etc/passwd','FS'));